braj_correction_agent/requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
python-docx>=1.2.0
python-dotenv>=1.0.0
